refactor(moderation): rename ReportError to ModerationError, move to error.rs
