+++ server/src/moderation/admin_handlers.rs
@@ -4,8 +4,9 @@ use axum::extract::{Path, Query, State};
 use axum::{Extension, Json};
 use uuid::Uuid;
 
+use super::error::ModerationError;
 use super::types::{
-    ListReportsQuery, PaginatedReports, Report, ReportError, ReportResponse, ReportStatsResponse,
+    ListReportsQuery, PaginatedReports, Report, ReportResponse, ReportStatsResponse,
     ResolveReportRequest,
 };
 use crate::admin::ElevatedAdmin;
@@ -25,7 +26,7 @@ use crate::ws::{broadcast_admin_event, ServerEvent};
 pub async fn list_reports(
     State(state): State<AppState>,
     Query(query): Query<ListReportsQuery>,
-) -> Result<Json<PaginatedReports>, ReportError> {
+) -> Result<Json<PaginatedReports>, ModerationError> {
     let limit = query.limit.clamp(1, 100);
     let offset = query.offset.max(0);
 
@@ -75,12 +76,12 @@ pub async fn list_reports(
 pub async fn get_report(
     State(state): State<AppState>,
     Path(report_id): Path<Uuid>,
-) -> Result<Json<ReportResponse>, ReportError> {
+) -> Result<Json<ReportResponse>, ModerationError> {
     let report = sqlx::query_as::<_, Report>("SELECT * FROM user_reports WHERE id = $1")
         .bind(report_id)
         .fetch_optional(&state.db)
         .await?
-        .ok_or(ReportError::NotFound)?;
+        .ok_or(ModerationError::NotFound)?;
 
     Ok(Json(report.into()))
 }
@@ -99,7 +100,7 @@ pub async fn claim_report(
     State(state): State<AppState>,
     Extension(elevated): Extension<ElevatedAdmin>,
     Path(report_id): Path<Uuid>,
-) -> Result<Json<ReportResponse>, ReportError> {
+) -> Result<Json<ReportResponse>, ModerationError> {
     let report = sqlx::query_as::<_, Report>(
         r"UPDATE user_reports
            SET status = 'reviewing', assigned_admin_id = $2, updated_at = NOW()
@@ -110,7 +111,7 @@ pub async fn claim_report(
     .bind(elevated.user_id)
     .fetch_optional(&state.db)
     .await?
-    .ok_or(ReportError::NotFound)?;
+    .ok_or(ModerationError::NotFound)?;
 
     Ok(Json(report.into()))
 }
@@ -130,11 +131,11 @@ pub async fn resolve_report(
     State(state): State<AppState>,
     Path(report_id): Path<Uuid>,
     Json(body): Json<ResolveReportRequest>,
-) -> Result<Json<ReportResponse>, ReportError> {
+) -> Result<Json<ReportResponse>, ModerationError> {
     // Validate resolution_action
     let valid_actions = ["dismissed", "warned", "banned", "escalated"];
     if !valid_actions.contains(&body.resolution_action.as_str()) {
-        return Err(ReportError::Validation(format!(
+        return Err(ModerationError::Validation(format!(
             "Invalid resolution action. Must be one of: {}",
             valid_actions.join(", ")
         )));
@@ -155,7 +156,7 @@ pub async fn resolve_report(
     .bind(&body.resolution_note)
     .fetch_optional(&state.db)
     .await?
-    .ok_or(ReportError::NotFound)?;
+    .ok_or(ModerationError::NotFound)?;
 
     // Broadcast resolution to admin events
     let event = ServerEvent::AdminReportResolved {
@@ -179,7 +180,7 @@ pub async fn resolve_report(
 )]
 pub async fn report_stats(
     State(state): State<AppState>,
-) -> Result<Json<ReportStatsResponse>, ReportError> {
+) -> Result<Json<ReportStatsResponse>, ModerationError> {
     let pending: i64 = sqlx::query_scalar::<_, Option<i64>>(
         "SELECT COUNT(*) FROM user_reports WHERE status ='pending'",
     )
