+++ server/src/moderation/mod.rs
@@ -4,6 +4,7 @@
 
 pub mod admin_handlers;
 pub mod defaults;
+pub mod error;
 pub mod filter_cache;
 pub mod filter_engine;
 pub mod filter_handlers;
@@ -11,3 +12,5 @@ pub mod filter_queries;
 pub mod filter_types;
 pub mod handlers;
 pub mod types;
+
+pub use error::ModerationError;
